Rename email module
